Lots of changes

+++ src/sorter.py
@@ -19,12 +19,11 @@ dest_dir = ""
 dest_structure = {}
 dir_contents = []
 
-def sort(src,dest,dir_list):
+def sort(src,dest,dir_list,debug=False):
     years = []
     im = ""
-
     to_from = {}
-    FAILURES = 0
+    new_dir = {}
 
     for x in dir_list: 
         im = Image.open(os.path.join(src,x))
@@ -38,36 +37,58 @@ def sort(src,dest,dir_list):
 
         month = datetime.datetime.fromtimestamp(t).strftime("%m")
         year = datetime.datetime.fromtimestamp(t).strftime("%y")
-        name = "20" + year    
+        name = "20" + year  
+        month_name = f"{month} - {MONTHS[month]}"  
         dest_dir = os.path.join(dest,name)
-        final_dest = os.path.join(dest_dir,MONTHS[month])
-
-        # print(src,final_dest)
+        final_dest = os.path.join(dest_dir,month_name)
 
         if not os.path.exists(final_dest):
-            os.makedirs(final_dest)
+            if debug:
+                new_dir[final_dest] = False
+            else:       
+                new_dir[final_dest] = True    
+                os.makedirs(final_dest)
+        
 
-        # print("MOVING:",os.path.join(src,x),"to",os.path.join(final_dest,x))
         to_from[os.path.join(src,x)] = os.path.join(final_dest,x)
-        # shutil.move(os.path.join(src,x),os.path.join(final_dest,x))
+    if debug:
+        print("--- NEW DIRECTORIES ---")
+        for x in new_dir:
+            print(f"CREATING New Directory:\t{x}")
+        print("--- MOVING FILES ---")
     for k,v in to_from.items():
-        move(k,v)
+        if debug:
+            print(f"MOVING: {k} \nTO:\t{v}")
+        else:
+            move(k,v)
 
 
-def unsort(src,dst,files={},dirs={}):
+def unsort(src,dst,files={},dirs=[],debug=False):
+
+    print(f"\n !!! Running unsort() in DEBUG mode with \nSRC: {src} \nDST: {dst}\nFFILES: {files}\nDIRECTORIES: {dirs}\n") if debug else None
+
     buffer = os.listdir(src)
     dirs = []
+    
     for x in buffer:
         if os.path.isfile(os.path.join(src,x)):
-            move(os.path.join(src,x),os.path.join(dst,x))
+            if debug:
+                print(f"MOVING: {x} \nTO:\t{os.path.join(dst,x)}")
+            else:
+                move(os.path.join(src,x),os.path.join(dst,x))
         else:
             dirs.append(os.path.join(src,x))
-            unsort(os.path.join(src,x),dst)
+            unsort(os.path.join(src,x),dst,files,dirs,debug)
+
     for x in dirs:
-        try:
-            os.rmdir(x)
-        except:
-            print("FAILED TO REMOVE:",x)
+        if debug:
+            print(f"REMOVING: {x}")
+        else:
+            try:
+                os.rmdir(x)
+            except:
+                print("FAILED TO REMOVE:",x)
+
 
 
 def move(src,dst):
